+++ extremitypathfinder/types.py
@@ -9,5 +9,6 @@ Length = Optional[float]
 InputNumerical = Union[float, int]
 InputCoord = Tuple[InputNumerical, InputNumerical]
 InputCoordList = Union[np.ndarray, List[InputCoord]]
+OptionalInputCoordList = Optional[InputCoordList]
 ObstacleIterator = Iterable[InputCoord]
 Graph = nx.Graph